initial commit of account_db functionality to branch
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/ApiCommonModel/branches/account_db@81738 735e2a04-f8fc-0310-8a1b-f2942603c481

+++ Model/src/java/org/apidb/apicommon/model/userdataset/GeneListTypeHandler.java
@@ -41,7 +41,7 @@ public class GeneListTypeHandler extends UserDatasetTypeHandler {
   }
 
   @Override
-  public String[] getUninstallInAppDbCommand(Integer userDatasetId, String projectId) {
+  public String[] getUninstallInAppDbCommand(Long userDatasetId, String projectId) {
     String[] cmd = {"uninstallGeneListUserDataset", userDatasetId.toString(), projectId};
     return cmd;
   }
